+++ book.rb
@@ -3,13 +3,6 @@ class Book
         @title = title
         @author = author
     end
-    def create_book
-        puts "Title:"
-        title = gets.chomp
-        puts "Author:"
-        author = gets.chomp
-        puts "Book created successfuly!"
-        @books << Book.new(title, author)
-    end
-
+   
+    attr_accessor :title, :author
 end
