+++ main.rb
@@ -42,15 +42,7 @@ def main
     type = gets.chomp
 
     case type
-    when 'teacher'
-    print 'Age: '
-      age = gets.chomp.to_i
-      print 'Name: '
-      name = gets.chomp
-      print 'Specialization:  '
-      specialization = gets.chomp
-      @people << Teacher.new(age ,name, specialization)
-    when 'student'
+    when 1
       print 'Age: '
       age = gets.chomp.to_i
       print 'Name: '
@@ -58,6 +50,14 @@ def main
       print 'Has parent permission? [Y/N]:  '
       permission = gets.chomp.downcase
       @people << Student.new(age, name, permission)
+    when 2
+    print 'Age: '
+      age = gets.chomp.to_i
+      print 'Name: '
+      name = gets.chomp
+      print 'Specialization:  '
+      specialization = gets.chomp
+      @people << Teacher.new(age ,name, specialization)
     else
       puts 'Invalid input, person not created'
     end
